+++ lib/nodekey/NodeKey.ts
@@ -1,7 +1,7 @@
 import CryptoJS from 'crypto-js';
-import { randomBytes } from 'crypto';
 import secp256k1 from 'secp256k1';
-import fs from 'fs';
+import { randomBytes } from '../utils/utils';
+import { exists, readFile, writeFile } from '../utils/fsUtils';
 
 /**
  * A class representing an ECDSA public/private key pair that identifies an XU node on the network
@@ -20,12 +20,12 @@ class NodeKey {
   }
 
   /**
-   * Generate a random NodeKey.
+   * Generates a random NodeKey.
    */
-  private static generate = (): NodeKey => {
+  private static generate = async (): Promise<NodeKey> => {
     let privKey: Buffer;
     do {
-      privKey = randomBytes(32);
+      privKey = await randomBytes(32);
     } while (!secp256k1.privateKeyVerify(privKey));
 
     return new NodeKey(privKey);
@@ -37,43 +37,44 @@ class NodeKey {
    * @param password an optional password to decrypt the file
    * @returns a NodeKey if a file containing a valid ECDSA private key exists at the given path
    */
-  private static fromFile = (path: string, password?: string): NodeKey => {
-    let buf: Buffer;
-    if (password) {
-      const encryptedString: string = fs.readFileSync(path, 'utf8');
-      const decryptedString: string = CryptoJS.AES.decrypt(encryptedString, password).toString(CryptoJS.enc.Hex);
-      buf = Buffer.from(decryptedString, 'hex');
-    } else {
-      buf = fs.readFileSync(path);
-    }
+  private static fromFile = (path: string, password?: string): Promise<NodeKey> => {
+    return new Promise(async (resolve, reject) => {
+      let buf = await readFile(path);
 
-    if (secp256k1.privateKeyVerify(buf)) {
-      return new NodeKey(buf);
-    } else {
-      throw new Error(`${path} does not contain a valid ECDSA private key`);
-    }
+      if (password) {
+        const encryptedString = buf.toString('utf8');
+        const decryptedString = CryptoJS.AES.decrypt(encryptedString, password).toString(CryptoJS.enc.Hex);
+        buf = Buffer.from(decryptedString, 'hex');
+      }
+
+      if (secp256k1.privateKeyVerify(buf)) {
+        resolve(new NodeKey(buf));
+      } else {
+        reject(new Error(`${path} does not contain a valid ECDSA private key`));
+      }
+    });
   }
 
   /**
-   * Load a node key from a file or create one if none exists. See [[fromFile]] and [[generate]].
+   * Loads a node key from a file or create one if none exists. See [[fromFile]] and [[generate]].
    */
-  public static load = (xudir: string, instanceId = 0): NodeKey => {
+  public static load = async (xudir: string, instanceId = 0): Promise<NodeKey> => {
     const path: string = instanceId > 0
       ? `${xudir}/nodekey_${instanceId}.dat`
       : `${xudir}/nodekey.dat`;
 
     let nodeKey: NodeKey;
-    if (fs.existsSync(path)) {
-      nodeKey = NodeKey.fromFile(path);
+    if (await exists(path)) {
+      nodeKey = await NodeKey.fromFile(path);
     } else {
-      nodeKey = NodeKey.generate();
-      nodeKey.toFile(path);
+      nodeKey = await NodeKey.generate();
+      await nodeKey.toFile(path);
     }
     return nodeKey;
   }
 
   /**
-   * Sign a message with the private key.
+   * Signs a message with the private key.
    * @param msg the data to sign
    * @returns the signature
    */
@@ -82,11 +83,11 @@ class NodeKey {
   }
 
   /**
-   * Save the private key to a file, optionally encrypted by a password.
+   * Saves the private key to a file, optionally encrypted by a password.
    * @param path the path at which to save the file
    * @param password an optional password parameter for encrypting the private key
    */
-  private toFile = (path: string, password?: string): void => {
+  private toFile = (path: string, password?: string): Promise<void> => {
     let buf: Buffer | CryptoJS.WordArray;
     if (password) {
       const lwa = CryptoJS.lib.WordArray.create(this.privKey.buffer);
@@ -94,7 +95,7 @@ class NodeKey {
     } else {
       buf = this.privKey;
     }
-    fs.writeFileSync(path, buf);
+    return writeFile(path, buf);
   }
 }