+++ lib/p2p/NodeList.ts
@@ -3,23 +3,26 @@ import P2PRepository from './P2PRepository';
 import { NodeInstance, NodeFactory } from '../types/db';
 import { Address } from '../types/p2p';
 import addressUtils from '../utils/addressUtils';
-import { UriParts } from '../utils/utils';
 import { ReputationEvent } from '../types/enums';
+import { db } from '../types';
 
-const reputationEventWeight = {
+export const reputationEventWeight = {
   [ReputationEvent.ManualBan]: Number.NEGATIVE_INFINITY,
   [ReputationEvent.ManualUnban]: 0,
   [ReputationEvent.PacketTimeout]: -1,
   [ReputationEvent.SwapFailure]: -10,
   [ReputationEvent.SwapSuccess]: 1,
+  [ReputationEvent.InvalidMessage]: -10,
+  [ReputationEvent.UnknownPacketType]: -20,
+  [ReputationEvent.UnparseableMessage]: -10,
 };
 
 // TODO: inform node about getting banned
 // TODO: remove reputation events after certain amount of time
 
 interface NodeList {
-  on(event: 'node.ban', listener: (nodePubKey: string) => void): this;
-  emit(event: 'node.ban', nodePubKey: string): boolean;
+  on(event: 'node.ban', listener: (nodePubKey: string, events: db.ReputationEventInstance[]) => void): this;
+  emit(event: 'node.ban', nodePubKey: string, events: db.ReputationEventInstance[]): boolean;
 }
 
 /** Represents a list of nodes for managing network peers activity */
@@ -141,7 +144,9 @@ class NodeList extends EventEmitter {
         promises.push(this.setBanStatus(node, true));
         this.bannedNodes.add(node.nodePubKey);
         this.nodes.delete(node.nodePubKey);
-        this.emit('node.ban', nodePubKey);
+
+        const events = await this.repository.getReputationEvents(node);
+        this.emit('node.ban', nodePubKey, events);
       } else if (node.banned) {
         // If the reputationScore is not below the banThreshold but node.banned
         // is true that means that the node was unbanned