+++ lib/p2p/Pool.ts
@@ -2,19 +2,20 @@ import net, { Server, Socket } from 'net';
 import { EventEmitter } from 'events';
 import errors, { errorCodes } from './errors';
 import Peer, { PeerInfo } from './Peer';
-import NodeList from './NodeList';
-import PeerList from './PeerList';
+import NodeList, { reputationEventWeight } from './NodeList';
 import P2PRepository from './P2PRepository';
 import * as packets from './packets/types';
 import { Packet, PacketType } from './packets';
-import { OutgoingOrder, OrderPortion, PeerOrder, IncomingOrder } from '../types/orders';
+import { OutgoingOrder, OrderPortion, IncomingOrder } from '../types/orders';
 import { Models } from '../db/DB';
 import Logger from '../Logger';
 import { HandshakeState, Address, NodeConnectionInfo, HandshakeStateUpdate } from '../types/p2p';
 import addressUtils from '../utils/addressUtils';
 import { getExternalIp, ms } from '../utils/utils';
 import assert from 'assert';
-import { ReputationEvent } from '../types/enums';
+import { ReputationEvent, DisconnectionReason } from '../types/enums';
+import { DisconnectingPacketBody } from './packets/types/DisconnectingPacket';
+import { db } from '../types';
 
 type PoolConfig = {
   /** Whether or not to automatically detect and share current external ip address on startup. */
@@ -39,7 +40,7 @@ type PoolConfig = {
 
 type NodeReputationInfo = {
   reputationScore: ReputationEvent;
-  banned: boolean;
+  banned?: boolean;
 };
 
 interface Pool {
@@ -70,12 +71,12 @@ interface NodeConnectionIterator {
 class Pool extends EventEmitter {
   /** The local handshake data to be sent to newly connected peers. */
   public handshakeData!: HandshakeState;
-  /** A set of pub keys of nodes for which we have pending outgoing connections. */
+  /** A map of pub keys to nodes for which we have pending outgoing connections. */
   private pendingOutgoingConnections = new Map<string, Peer>();
   /** A collection of known nodes on the XU network. */
   private nodes: NodeList;
   /** A collection of opened, active peers. */
-  private peers: PeerList = new PeerList();
+  private peers = new Map<string, Peer>();
   private server?: Server;
   private connected = false;
   /** The port on which to listen for peer connections, undefined if this node is not listening. */
@@ -115,7 +116,7 @@ class Pool extends EventEmitter {
   }
 
   public get peerCount(): number {
-    return this.peers.length;
+    return this.peers.size;
   }
 
   /**
@@ -200,12 +201,13 @@ class Pool extends EventEmitter {
   }
 
   private bindNodeList = () => {
-    this.nodes.on('node.ban', (nodePubKey) => {
+    this.nodes.on('node.ban', (nodePubKey: string, events: db.ReputationEventInstance[]) => {
       this.logger.warn(`node ${nodePubKey} was banned`);
 
       const peer = this.peers.get(nodePubKey);
       if (peer) {
-        peer.close();
+        const lastNegativeEvents = events.filter(e => reputationEventWeight[e.event] < 0).slice(0, 10);
+        peer.close({ reason: DisconnectionReason.Banned, payload: JSON.stringify(lastNegativeEvents) });
       }
     });
   }
@@ -345,7 +347,7 @@ class Pool extends EventEmitter {
   }
 
   public listPeers = (): PeerInfo[] => {
-    const peerInfos: PeerInfo[] = Array.from({ length: this.peers.length });
+    const peerInfos: PeerInfo[] = Array.from({ length: this.peers.size });
     let i = 0;
     this.peers.forEach((peer) => {
       peerInfos[i] = peer.info;
@@ -381,10 +383,10 @@ class Pool extends EventEmitter {
     }
   }
 
-  public closePeer = async (nodePubKey: string): Promise<void> => {
+  public closePeer = async (nodePubKey: string, reason?: DisconnectingPacketBody): Promise<void> => {
     const peer = this.peers.get(nodePubKey);
     if (peer) {
-      peer.close();
+      peer.close(reason);
       this.logger.info(`Disconnected from ${peer.nodePubKey}@${addressUtils.toString(peer.address)}`);
     } else {
       throw(errors.NOT_CONNECTED(nodePubKey));
@@ -500,7 +502,7 @@ class Pool extends EventEmitter {
       }
       case PacketType.OrderInvalidation: {
         const orderPortion = (packet as packets.OrderInvalidationPacket).body!;
-        this.logger.verbose(`canceled order from ${peer.nodePubKey}: ${JSON.stringify(orderPortion)}`);
+        this.logger.verbose(`received order invalidation from ${peer.nodePubKey}: ${JSON.stringify(orderPortion)}`);
         this.emit('packet.orderInvalidation', orderPortion, peer.nodePubKey as string);
         break;
       }
@@ -554,28 +556,33 @@ class Pool extends EventEmitter {
         this.emit('packet.swapError', packet);
         break;
       }
+
+      case PacketType.Disconnecting: {
+        this.logger.debug(`received disconnecting packet from ${peer.nodePubKey}: ${JSON.stringify(packet.body)}`);
+        break;
+      }
     }
   }
 
   private handleOpen = async (peer: Peer): Promise<void> => {
     if (!this.connected) {
       // if we have disconnected the pool, don't allow any new connections to open
-      peer.close();
+      peer.close({ reason: DisconnectionReason.NotAcceptingConnections });
       return;
     }
-    if (peer.nodePubKey === this.handshakeData.nodePubKey) {
+    if (!peer.nodePubKey || peer.nodePubKey === this.handshakeData.nodePubKey) {
       return;
     }
 
-    if (this.nodes.isBanned(peer.nodePubKey!)) {
+    if (this.nodes.isBanned(peer.nodePubKey)) {
       // TODO: Ban IP address for this session if banned peer attempts repeated connections.
-      peer.close();
-    } else if (this.peers.has(peer.nodePubKey!)) {
+      peer.close({ reason: DisconnectionReason.Banned });
+    } else if (this.peers.has(peer.nodePubKey)) {
       // TODO: Penalize peers that attempt to create duplicate connections to us
-      peer.close();
+      peer.close({ reason: DisconnectionReason.AlreadyConnected });
     } else {
       this.logger.verbose(`opened connection to ${peer.nodePubKey} at ${addressUtils.toString(peer.address)}`);
-      this.peers.add(peer);
+      this.peers.set(peer.nodePubKey, peer);
 
       // request peer's orders
       peer.sendPacket(new packets.GetOrdersPacket({ pairIds: this.handshakeData.pairs }));
@@ -594,15 +601,15 @@ class Pool extends EventEmitter {
       });
 
       // upserting the node entry
-      if (!this.nodes.has(peer.nodePubKey!)) {
+      if (!this.nodes.has(peer.nodePubKey)) {
         await this.nodes.createNode({
           addresses,
-          nodePubKey: peer.nodePubKey!,
+          nodePubKey: peer.nodePubKey,
           lastAddress: peer.inbound ? undefined : peer.address,
         });
       } else {
         // the node is known, update its listening addresses
-        await this.nodes.updateAddresses(peer.nodePubKey!, addresses, peer.inbound ? undefined : peer.address);
+        await this.nodes.updateAddresses(peer.nodePubKey, addresses, peer.inbound ? undefined : peer.address);
       }
     }
   }
@@ -655,24 +662,21 @@ class Pool extends EventEmitter {
     peer.once('close', () => {
       if (peer.nodePubKey) {
         this.pendingOutgoingConnections.delete(peer.nodePubKey);
-        this.peers.remove(peer.nodePubKey);
+        this.peers.delete(peer.nodePubKey);
       }
       this.emit('peer.close', peer);
     });
 
-    peer.once('ban', async () => {
-      this.logger.debug(`Banning peer (${peer.nodePubKey})`);
+    peer.once('reputation', async (event) => {
+      this.logger.debug(`Peer (${peer.nodePubKey}), received reputation event: ${ReputationEvent[event]}`);
       if (peer.nodePubKey) {
-        await this.nodes.ban(peer.nodePubKey);
-      }
-      if (peer.connected) {
-        peer.close();
+        await this.nodes.addReputationEvent(peer.nodePubKey, event);
       }
     });
   }
 
   private closePeers = () => {
-    this.peers.forEach(peer => peer.close());
+    this.peers.forEach(peer => peer.close({ reason: DisconnectionReason.Shutdown }));
   }
 
   private closePendingConnections = () => {