+++ lib/types/enums.ts
@@ -47,6 +47,9 @@ export enum ReputationEvent {
   PacketTimeout = 2,
   SwapFailure = 3,
   SwapSuccess = 4,
+  UnparseableMessage = 5,
+  InvalidMessage = 6,
+  UnknownPacketType = 7,
 }
 
 export enum SwapFailureReason {
@@ -56,4 +59,16 @@ export enum SwapFailureReason {
   OrderOnHold,
   /** The swap request contained invalid data. */
   InvalidSwapRequest,
+}
+
+export enum DisconnectionReason {
+  ResponseStalling,
+  IncompatibleProtocolVersion,
+  UnexpectedIdentity,
+  ForbiddenIdentityUpdate,
+  ConnectedToSelf,
+  NotAcceptingConnections,
+  Banned,
+  AlreadyConnected,
+  Shutdown,
 }